Fix new transfer protocol for LLVM compiler
git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS/dev/dev_ym/XIOS_COUPLING@2347 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/transport/one_sided_client_buffer.cpp
@@ -117,7 +117,7 @@ namespace xios
         currentBuffer->write(&buffer, size, addr, start, count) ;
         if (count > 0) 
         {
-          blocs_.push_back({addr,currentBuffer_, start, count, currentWindow_}) ;
+          blocs_.push_back({addr,currentBuffer_, start, static_cast<int>(count), currentWindow_}) ;
           nbBlocs++ ; 
         }
 
@@ -126,7 +126,7 @@ namespace xios
 
         if (count > 0) 
         {
-          blocs_.push_back({addr,currentBuffer_, start, count, currentWindow_}) ;
+          blocs_.push_back({addr,currentBuffer_, start, static_cast<int>(count), currentWindow_}) ;
           nbBlocs++ ; 
         }
